feat: introduce helper package for implementing a custom JSDOM environment
