+++ packages/jest-environment-jsdom-abstract/.npmignore
@@ -1,0 +1,8 @@
+**/__mocks__/**
+**/__tests__/**
+__typetests__
+src
+tsconfig.json
+tsconfig.tsbuildinfo
+api-extractor.json
+.eslintcache
