+++ packages/jest-environment-jsdom-abstract/tsconfig.json
@@ -8,6 +8,9 @@
   "exclude": ["./**/__mocks__/**/*", "./**/__tests__/**/*"],
   "references": [
     {"path": "../jest-environment"},
-    {"path": "../jest-environment-jsdom-abstract"}
+    {"path": "../jest-fake-timers"},
+    {"path": "../jest-mock"},
+    {"path": "../jest-types"},
+    {"path": "../jest-util"}
   ]
 }
